Points Implemented
The cities now have points! Added my filesystem to the config, left a
description for protocol in comments, should be pretty easy to use.
Have NOT implemented coverage data yet, but soon... SOON.
