+++ etc/cities.cfg
@@ -1,13 +1,23 @@
 # Cities configuration file
 
+#MAP FORMAT
+#	.shp
+#		=path
+#	.csv
+#		=path latitudecolindex longitudecolindex
+#	.kml
+#		=path tagname
 
 [Washington DC]
-bus=dc_data/DCBudgetMetroBusOrignal.csv
 metro=dc_data/DCBudgetMetroRailOrignal.csv
+bus=dc_data/DCBudgetMetroBusOrignal.csv
+metromap=dc_data/dc_metro_stations/Metro_Stations_Regional.shp
+busmap=dc_data/dc_bus_stations/Metro_Bus_Stops.shp
 
 
 
 [New York City]
-bus=nyc_data/nyc_bus.csv
 metro=nyc_data/nyc_metro.csv
-
+bus=nyc_data/nyc_bus.csv
+metromap=nyc_data/nyc_metro/Stations.csv 9 10
+busmap=nyc_data/nyc_bus/BusStopsAsOfMarch2.shp
